[remake] TimerSystem and Add TimerOutQueue
全体のタイマー値を管理するGlobalTimerとタイムアウトキューを管理するLocalTimerに分離した。

動作確認をしていない

+++ src/kernel/task_manager/scheduling_class.rs
@@ -15,8 +15,8 @@ impl SchedulingClass {
         &self,
         priority_level: u8,
         number_of_threads: usize,
-        interval_ms: usize,
-    ) -> usize {
+        interval_ms: u64,
+    ) -> u64 {
         match self {
             SchedulingClass::KernelThread(s) => {
                 s.calculate_time_slice(priority_level, number_of_threads, interval_ms)
